requirements.txt
# Backend
fastapi==0.100.0
uvicorn[standard]
pydantic
Jinja2
aiofiles

# Testing
pytest
requests
allure-pytest
pytest-playwright

selenium
webdriver-manager

# Reporting
allure-combine